cryptomarket/public/serializers.py
from rest_framework import serializers
from order.models import Instrument, Transaction
from order.serializers import (
    TransactionSerializer,
    InstrumentSerializer,
    L2OrderBookSerializer,
    LevelSerializer
)

# Все сериализаторы уже определены в order.serializers, мы просто импортируем их 