pyproject.toml
[project]
name = "cryptomarket"
version = "0.1.0"
description = ""
authors = [
    {name = "stoicismguy",email = "pyzan55@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
